css changes to be more in style with new theme made by anmol
