+++ frontend/src/components/7.Result/ResultPage.css
@@ -1,6 +1,6 @@
 .ResultTitle {
   color: white;
-  font-family: Comic Sans MS;
+  font-family:monospace;
   display: flex;
   flex-direction: column;
   align-items: center;
